test(runtime): isolate AERROR innermost-line regression
